chore(metrics): Added presigned url metric
